% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_waic}
\alias{calculate_waic}
\title{Calculate WAIC from partially or weakly separable model}
\usage{
calculate_waic(result)
}
\description{
Calculate WAIC from partially or weakly separable model
}
