\name{rbfda-package}
\alias{rbfda-package}
\alias{rbfda}
\docType{package}
\title{
\packageTitle{rbfda}
}
\description{
\packageDescription{rbfda}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rbfda}
\packageIndices{rbfda}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rbfda}

Maintainer: \packageMaintainer{rbfda}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
